refactor(worker): split `sealing_config` and `sealing_store`
