+++ venus-worker/src/sealing/sealing_thread/mod.rs
@@ -1,11 +1,16 @@
+use std::collections::HashSet;
+use std::path::PathBuf;
 use std::time::{Duration, Instant};
 
 use anyhow::{Context, Result};
 use crossbeam_channel::{select, TryRecvError};
 
+use crate::config::{Sealing, SealingOptional};
 use crate::logging::{error, info, warn};
+use crate::store::Location;
 use crate::watchdog::{Ctx, Module};
 
+use super::config::{merge_sealing_fields, Config};
 use super::{failure::*, store::Store};
 
 mod task;
@@ -15,25 +20,40 @@ mod ctrl;
 pub use ctrl::Ctrl;
 use ctrl::*;
 
-pub struct Worker {
+pub struct SealingThread {
     idx: usize,
-    store: Store,
+
+    /// the config of this SealingThread
+    pub config: Config,
+    pub store: Store,
+
     ctrl_ctx: CtrlCtx,
 }
 
-impl Worker {
-    pub fn new(idx: usize, s: Store) -> (Self, Ctrl) {
-        let (ctrl, ctrl_ctx) = new_ctrl(s.location.clone());
-        (Worker { idx, store: s, ctrl_ctx }, ctrl)
+impl SealingThread {
+    pub fn new(idx: usize, plan: Option<String>, sealing_config: Sealing, location: Location) -> Result<(Self, Ctrl)> {
+        let store_path = location.to_pathbuf();
+        let store = Store::open(store_path).with_context(|| format!("open store {}", location.as_ref().display()))?;
+        let (ctrl, ctrl_ctx) = new_ctrl(location.clone());
+
+        Ok((
+            Self {
+                config: Config::new(sealing_config, plan, location.hot_config_path())?,
+                store,
+                ctrl_ctx,
+                idx,
+            },
+            ctrl,
+        ))
     }
 
     fn seal_one(&mut self, ctx: &Ctx, state: Option<State>) -> Result<(), Failure> {
-        let task = Task::build(ctx, &self.ctrl_ctx, &mut self.store)?;
+        let task = Task::build(ctx, &self.ctrl_ctx, &mut self.config, &mut self.store)?;
         task.exec(state)
     }
 }
 
-impl Module for Worker {
+impl Module for SealingThread {
     fn should_wait(&self) -> bool {
         false
     }
@@ -111,7 +131,7 @@ impl Module for Worker {
             }
 
             info!(
-                duration = ?self.store.config.seal_interval,
+                duration = ?self.config.seal_interval,
                 "wait before sealing"
             );
 
@@ -125,10 +145,47 @@ impl Module for Worker {
                     return Ok(())
                 },
 
-                default(self.store.config.seal_interval) => {
+                default(self.config.seal_interval) => {
 
                 }
             }
         }
+    }
+}
+
+pub(crate) fn build_sealing_threads(
+    list: &[crate::config::SealingThread],
+    common: &SealingOptional,
+) -> Result<Vec<(SealingThread, (usize, Ctrl))>> {
+    let mut sealing_threads = Vec::new();
+    let mut path_set = HashSet::new();
+
+    for (idx, scfg) in list.iter().enumerate() {
+        let sealing_config = customized_sealing_config(common, scfg.inner.sealing.as_ref());
+        let plan = scfg.inner.plan.as_ref().cloned();
+
+        let store_path = PathBuf::from(&scfg.location)
+            .canonicalize()
+            .with_context(|| format!("canonicalize store path {}", &scfg.location))?;
+        if path_set.contains(&store_path) {
+            tracing::warn!(path = ?store_path, "store already loaded");
+            continue;
+        }
+
+        let (sealing_thread, ctrl) = SealingThread::new(idx, plan, sealing_config, Location::new(store_path.clone()))?;
+        sealing_threads.push((sealing_thread, (idx, ctrl)));
+        path_set.insert(store_path);
+    }
+
+    Ok(sealing_threads)
+}
+
+fn customized_sealing_config(common: &SealingOptional, customized: Option<&SealingOptional>) -> Sealing {
+    let default_sealing = Sealing::default();
+    let common_sealing = merge_sealing_fields(default_sealing, common.clone());
+    if let Some(customized) = customized.cloned() {
+        merge_sealing_fields(common_sealing, customized)
+    } else {
+        common_sealing
     }
 }
