+++ venus-worker/src/sealing/sealing_thread/task/planner/common.rs
@@ -418,7 +418,7 @@ pub fn persist_sector_files(task: &'_ Task<'_>, cache_dir: Entry, sealed_file: E
             candidates=?candidates,
             "no persist store selected, wait for next polling"
         );
-        task.wait_or_interrupted(task.store.config.rpc_polling_interval)?;
+        task.wait_or_interrupted(task.sealing_config.rpc_polling_interval)?;
     };
 
     let persist_store = task
