+++ venus-worker/src/sealing/sealing_thread/task/planner/rebuild.rs
@@ -124,8 +124,8 @@ impl<'c, 't> Rebuild<'c, 't> {
             self.task.ctx.global.rpc,
             allocate_rebuild_sector,
             AllocateSectorSpec {
-                allowed_miners: Some(self.task.store.config.allowed_miners.clone()),
-                allowed_proof_types: Some(self.task.store.config.allowed_proof_types.clone()),
+                allowed_miners: Some(self.task.sealing_config.allowed_miners.clone()),
+                allowed_proof_types: Some(self.task.sealing_config.allowed_proof_types.clone()),
             },
         };
